% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt_binning.R
\name{opt_binning}
\alias{opt_binning}
\title{Optimal binning of dataframe}
\usage{
opt_binning(data, target, exclude_variables = NULL, min_lvl = 8,
  p = 0.05)
}
\description{
This function returns a dataframe with all variables that
 have not been excluded in the "exclude_variables" list, and that have more than the min_lvl
 amount of levels, in optimal bins to the target.
}
\examples{
opt_binning()
}
\keyword{binning}
\keyword{binning,}
\keyword{optimal}
